requirements.txt
aiohttp==3.8.5
aiomcache==0.8.1
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
blinker==1.6.2
boto3==1.28.2
botocore==1.31.56
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==3.4.8
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-cache2==0.2.1
fastapi-jwt-auth==0.5.0
fastapi-limiter==0.1.5
fastapi-mail==1.2.8
fastapi-redis-cache==0.2.5
FastAPI-SQLAlchemy==0.2.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
passlib==1.7.4
pendulum==2.1.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
pytzdata==2020.1
razorpay==1.4.1
redis==4.5.5
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
websockets==11.0.3
yarl==1.9.2
